libswitchhw: return version as integer to ensure comparison

+++ userspace/libswitchhw/i2c_io.h
@@ -25,7 +25,7 @@ int shw_i2c_io_scan(uint8_t *dev_map);
 uint8_t shw_i2c_read(const *i2c_bus, uint8_t addr);
 uint8_t shw_i2c_write(const *i2c_bus, uint8_t addr, uint8_t value, uint8_t mask);
 
-float shw_get_hw_ver();
+int shw_get_hw_ver();
 uint8_t shw_get_fpga_type();
 
 #endif			//I2C_SFP_H
